Add functionality to create a game
Add api function to post a new game.
Adjust function to create a game so that it marries the user input with
the expected prepopulated fields, then makes the request.
Then added UI calls to dictate behavior on success and failure.
Now adding a 'get' request to the get games success.