Add functionality to index games
Add index games request that takes each game displays their stats in a
table.
This involved appending an index request to the successful creation of
a game, then making the UI append each individual game to the table.
Next will add functionality so that the index of games load upon
the click of the games page button.
After that will add delete functionality.